Rename back to pydeck-carto
