--- bindings/carto-pydeck/carto_pydeck/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.1.0b1"
