+++ bindings/pydeck-carto/examples/scripts/carto_layer_geo_query_param.py
@@ -2,21 +2,22 @@
 CartoLayer
 ==========
 
-Render cloud data from a query.
+Render cloud data with query parameters.
 """
 import pydeck as pdk
-
 from carto_auth import CartoAuth
-from carto_pydeck import register_carto_layer, get_layer_credentials, notify_error
-from carto_pydeck.layer import MapType, CartoConnection
-
-register_carto_layer()
+from pydeck_carto import register_carto_layer, get_layer_credentials, notify_error
+from pydeck_carto.layer import MapType, CartoConnection
 
 carto_auth = CartoAuth.from_oauth()
 
+register_carto_layer()
+
 layer = pdk.Layer(
     "CartoLayer",
-    data="SELECT geom, name FROM carto-demo-data.demo_tables.airports",
+    data="SELECT geom, event FROM carto-demo-data.demo_tables"
+    ".spain_earthquakes where depth > ?",
+    query_parameters=[2],
     type_=MapType.QUERY,
     connection=CartoConnection.CARTO_DW,
     credentials=get_layer_credentials(carto_auth),
@@ -26,7 +27,7 @@ layer = pdk.Layer(
     on_data_error=notify_error(),
 )
 
-view_state = pdk.ViewState(latitude=0, longitude=0, zoom=1)
+view_state = pdk.ViewState(latitude=36, longitude=-7.44, zoom=5)
 
 r = pdk.Deck(layer, map_style=pdk.map_styles.ROAD, initial_view_state=view_state)
-r.to_html("carto_layer_geo_query.html", open_browser=True)
+r.to_html("carto_layer_geo_query_param.html", open_browser=True)
