+++ bindings/pydeck-carto/examples/scripts/carto_styles_color_continuous.py
@@ -5,15 +5,14 @@ CartoLayer
 Render cloud data with color continuous style.
 """
 import pydeck as pdk
-
 from carto_auth import CartoAuth
-from carto_pydeck import register_carto_layer, get_layer_credentials, notify_error
-from carto_pydeck.layer import MapType, CartoConnection
-from carto_pydeck.styles import color_continuous
-
-register_carto_layer()
+from pydeck_carto import register_carto_layer, get_layer_credentials, notify_error
+from pydeck_carto.layer import MapType, CartoConnection
+from pydeck_carto.styles import color_continuous
 
 carto_auth = CartoAuth.from_oauth()
+
+register_carto_layer()
 
 layer = pdk.Layer(
     "CartoLayer",
