+++ bindings/pydeck-carto/pydeck_carto/_version.py
@@ -1,0 +1,1 @@
+__version__ = "0.1.0b6"
