+++ bindings/pydeck-carto/tests/test_styles.py
@@ -1,4 +1,4 @@
-from carto_pydeck.styles import (
+from pydeck_carto.styles import (
     color_bins,
     color_categories,
     color_continuous,
