+++ bindings/pydeck/setup.py
@@ -205,7 +205,7 @@ if __name__ == "__main__":
             "Framework :: Jupyter",
         ],
         extras_require={
-            "carto": ["carto-pydeck"],
+            "carto": ["pydeck-carto"],
             "jupyter": [
                 "ipywidgets>=7,<8",
                 "traitlets>=4.3.2",
